backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.20
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
cryptography==44.0.2
pandas==2.2.3
matplotlib==3.10.1
numpy==1.24.3
openpyxl==3.1.5
pymongo==4.11.2
motor==3.7.0
python-dotenv==1.0.0
httpx==0.24.1
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-cov==4.1.0
beautifulsoup4==4.12.2
redis==5.0.1